convert webpack config to .ts
