+++ apps/extension/package.json
@@ -5,9 +5,9 @@
   "description": "chrome-extension",
   "type": "module",
   "scripts": {
-    "dev": "webpack --watch --mode=development -d inline-source-map",
+    "dev": "pnpm build --watch --mode=development -d inline-source-map",
     "clean": "rm -rfv dist bin",
-    "build": "webpack",
+    "build": "NODE_OPTIONS=\"--import=./src/utils/webpack-register.js\" webpack",
     "lint": "eslint \"**/*.ts*\"",
     "test": "vitest run"
   },
@@ -50,6 +50,7 @@
     "@types/lodash": "^4.17.0",
     "@types/react": "^18.2.72",
     "@types/react-dom": "^18.2.22",
+    "@types/webpack": "^5.28.5",
     "autoprefixer": "^10.4.19",
     "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^6.10.0",
@@ -61,6 +62,7 @@
     "style-loader": "^3.3.4",
     "tailwindcss": "^3.4.1",
     "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.2",
     "vite-plugin-top-level-await": "^1.4.1",
     "vite-plugin-wasm": "^3.3.0",
     "webpack": "^5.91.0",
